requirements.txt
python-dotenv
langchain
langchain-openai
asteval
requests
